sslocal: create test file for reported sql stats tests
Move tests for reported sql stats into its own file for clarity.

Epic: none

Release note: None
